[Issue-1795] Rollback tab
